moved register job integration test for better cohesion
